Removed app theme in app pkg
