--- app/lib/app/data/source/app_local_data_source.dart
@@ -1,3 +1,0 @@
-abstract class AppLocalDataSource {
-  String get appVersion;
-}
