--- app/lib/app/data/source/theme_local_data_source.dart
@@ -1,7 +1,0 @@
-import 'package:my_quran/theme/theme.dart';
-
-abstract class ThemeLocalDataSource {
-  CustomTheme get initialTheme;
-  Future<void> saveThemePrimaryColor(int index);
-  Future<void> saveThemeMode({required bool isDark});
-}
