+++ app/lib/app/domain/domain.dart
@@ -1,10 +1,4 @@
-export 'repository/app_repositoty.dart';
-export 'repository/theme_repository.dart';
 export 'repository/auth_repository.dart';
-export 'usecase/get_app_version_use_case.dart';
-export 'usecase/get_initial_theme_use_case.dart';
-export 'usecase/set_color_use_case.dart';
-export 'usecase/set_mode_use_case.dart';
 export 'usecase/get_initial_user_use_case.dart';
 export 'usecase/google_sign_in_use_case.dart';
 export 'usecase/apple_sign_in_use_case.dart';
