--- app/lib/app/domain/repository/app_repositoty.dart
@@ -1,3 +1,0 @@
-abstract class AppRepository {
-  String get appVersion;
-}
