+++ app/lib/app/presentation/cubit/app_cubit.dart
@@ -1,29 +1,23 @@
 import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-
-import 'package:my_quran/app/app.dart';
-import 'package:my_quran/theme/custom/custom_theme.dart';
+import 'package:mq_app_theme/mq_app_theme.dart';
 
 part 'app_state.dart';
 
 class AppCubit extends Cubit<AppState> {
-  AppCubit({
-    required this.getInitialThemeUseCase,
-    required this.setModeUseCase,
-    required this.setColorUseCase,
-    required this.getAppVersionUseCase,
-  }) : super(
-          AppState(getInitialThemeUseCase.call, getAppVersionUseCase.call),
+  AppCubit(this.appThemeRepository)
+      : super(
+          AppState(
+            appThemeRepository.getInitialThemeState(),
+            appThemeRepository.appVersion(),
+          ),
         );
 
-  final GetAppInitialThemeUseCase getInitialThemeUseCase;
-  final SetModeUseCase setModeUseCase;
-  final SetColorUseCase setColorUseCase;
-  final GetAppVersionUseCase getAppVersionUseCase;
+  final AppRepository appThemeRepository;
 
   Future<void> changeMode({required bool isDark}) async {
-    await setModeUseCase(isDark: isDark);
+    await appThemeRepository.saveThemeMode(isDark: isDark);
     emit(
       state.copyWith(
         theme: state.theme.copyWith(
@@ -34,7 +28,7 @@ class AppCubit extends Cubit<AppState> {
   }
 
   Future<void> changeColor(int index, Color color) async {
-    await setColorUseCase(index, color);
+    await appThemeRepository.saveThemePrimaryColor(index);
     emit(state.copyWith(theme: state.theme.copyWith(targetColor: color)));
   }
 }
