+++ app/lib/modules/login/view/sign_in_view.dart
@@ -3,6 +3,7 @@ import 'dart:developer';
 import 'dart:io';
 
 import 'package:mq_analytics/mq_analytics.dart';
+import 'package:mq_app_theme/mq_app_theme.dart';
 import 'package:sign_in_with_apple/sign_in_with_apple.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -16,7 +17,6 @@ import 'package:my_quran/config/config.dart';
 import 'package:my_quran/constants/contants.dart';
 import 'package:my_quran/core/core.dart';
 import 'package:my_quran/l10n/l10.dart';
-import 'package:my_quran/theme/theme.dart';
 import 'package:my_quran/utils/urils.dart';
 
 class SignInView extends StatefulWidget {
