+++ app/lib/modules/login/view/verification_code_view.dart
@@ -2,11 +2,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:go_router/go_router.dart';
 import 'package:mq_analytics/mq_analytics.dart';
+import 'package:mq_app_theme/mq_app_theme.dart';
 import 'package:mq_ci_keys/mq_ci_keys.dart';
 import 'package:my_quran/app/presentation/presenation.dart';
 import 'package:my_quran/components/components.dart';
 import 'package:my_quran/config/config.dart';
-import 'package:my_quran/theme/theme.dart';
 import 'package:my_quran/utils/urils.dart';
 import 'package:pin_code_fields/pin_code_fields.dart';
