--- app/lib/theme/custom/color/app_colors.dart
@@ -1,13 +1,0 @@
-import 'package:flutter/material.dart';
-
-@immutable
-final class AppColors {
-  const AppColors._();
-
-  static const Color white = Colors.white;
-  static const Color black = Colors.black;
-  static const Color red = Colors.red;
-  static const Color yellow = Colors.yellow;
-  static const Color green = Colors.green;
-  static const Color greyBlack = Color.fromARGB(255, 28, 28, 28);
-}
