--- app/lib/theme/custom/typography/typography_theme.dart
@@ -1,50 +1,0 @@
-import 'package:flutter/material.dart';
-
-extension TypographyUtils on BuildContext {
-  ThemeData get theme => Theme.of(this);
-  TextTheme get textTheme => theme.textTheme;
-  ColorScheme get colors => theme.colorScheme;
-
-  TextStyle? get displayLarge => textTheme.displayLarge?.copyWith(
-        color: colors.onSurface,
-      );
-  TextStyle? get displayMedium => textTheme.displayMedium?.copyWith(
-        color: colors.onSurface,
-      );
-  TextStyle? get displaySmall => textTheme.displaySmall?.copyWith(
-        color: colors.onSurface,
-      );
-  TextStyle? get headlineLarge => textTheme.headlineLarge?.copyWith(
-        color: colors.onSurface,
-      );
-  TextStyle? get headlineMedium => textTheme.headlineMedium?.copyWith(
-        color: colors.onSurface,
-      );
-  TextStyle? get headlineSmall => textTheme.headlineSmall?.copyWith(
-        color: colors.onSurface,
-      );
-  TextStyle? get titleLarge => textTheme.titleLarge?.copyWith(
-        color: colors.onSurface,
-      );
-  TextStyle? get titleMedium => textTheme.titleMedium?.copyWith(
-        color: colors.onSurface,
-      );
-  TextStyle? get titleSmall => textTheme.titleSmall?.copyWith(
-        color: colors.onSurface,
-      );
-  TextStyle? get labelLarge => textTheme.labelLarge?.copyWith(
-        color: colors.onSurface,
-      );
-  TextStyle? get labelMedium => textTheme.labelMedium?.copyWith(
-        color: colors.onSurface,
-      );
-  TextStyle? get labelSmall => textTheme.labelSmall?.copyWith(
-        color: colors.onSurface,
-      );
-  TextStyle? get bodyLarge => textTheme.bodyLarge?.copyWith(
-        color: colors.onSurface,
-      );
-  TextStyle? get bodyMedium => textTheme.bodyMedium?.copyWith(
-        color: colors.onSurface,
-      );
-}
