--- app/lib/theme/theme.dart
@@ -1,5 +1,0 @@
-export 'custom/color/custom_color.dart';
-export 'custom/component/componet_theme.dart';
-export 'custom/custom_theme.dart';
-export 'custom/typography/typography_theme.dart';
-export 'custom/color/app_colors.dart';
